MNT #95 move the geometries
